Move GherkinDocumentWalker and pretty to gherkin-utils package
