+++ gherkin-utils/.rsync
@@ -1,0 +1,1 @@
+../.templates/default.mk default.mk
