+++ gherkin-utils/Makefile
@@ -1,0 +1,1 @@
+include default.mk
