+++ gherkin-utils/javascript/.gitignore
@@ -1,0 +1,18 @@
+dist/
+.idea/
+.nyc_output/
+coverage/
+node_modules/
+yarn.lock
+package-lock.json
+*.log
+.deps
+.tested*
+.linted
+.built*
+.compared
+.codegen
+acceptance/
+storybook-static
+*-go
+*.iml
