+++ gherkin-utils/javascript/.nycrc.json
@@ -1,0 +1,16 @@
+{
+  "extension": [
+    ".ts",
+    ".tsx"
+  ],
+  "exclude": [
+    "coverage",
+    "dist",
+    "src/**/*.d.ts"
+  ],
+  "reporter": [
+    "html",
+    "text"
+  ],
+  "all": true  
+}
