+++ gherkin-utils/javascript/.rsync
@@ -1,0 +1,3 @@
+../../LICENSE LICENSE
+../../.templates/github/ .github/
+../../.templates/javascript/ .
