+++ gherkin-utils/javascript/src/GherkinDocumentWalker.ts
