+++ gherkin-utils/javascript/src/index.ts
@@ -1,0 +1,4 @@
+import pretty from './pretty'
+import GherkinDocumentWalker from './GherkinDocumentWalker'
+
+export { pretty, GherkinDocumentWalker }
