+++ gherkin/javascript/src/index.ts
@@ -6,7 +6,6 @@ import Parser from './Parser'
 import AstBuilder from './AstBuilder'
 import TokenScanner from './TokenScanner'
 import compile from './pickles/compile'
-import pretty from './pretty/pretty'
 import DIALECTS from './gherkin-languages.json'
 import parse from '../test/walker/parse'
 
@@ -23,5 +22,4 @@ export {
   TokenScanner,
   compile,
   parse,
-  pretty,
 }
