+++ react/javascript/package.json
@@ -23,6 +23,7 @@
     "@cucumber/messages": "file:../../messages/javascript",
     "@cucumber/query": "file:../../query/javascript",
     "@cucumber/tag-expressions": "file:../../tag-expressions/javascript",
+    "@cucumber/gherkin-utils": "file:../../gherkin-utils/javascript",
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-regular-svg-icons": "^5.13.0",
