+++ react/javascript/test/search/TagSearchTest.ts
@@ -1,7 +1,8 @@
 import { messages } from '@cucumber/messages'
 import assert from 'assert'
 import { parseAndCompile } from './parse'
-import { Query as GherkinQuery, pretty } from '@cucumber/gherkin'
+import { Query as GherkinQuery } from '@cucumber/gherkin'
+import { pretty } from '@cucumber/gherkin-utils'
 import TagSearch from '../../src/search/TagSearch'
 
 describe('TagSearchTest', () => {
